migrate to numba, performance boost tensor_flow spike for loops

+++ test/tool/radprop/algorithms/avhrr_ndvi_test.py
@@ -15,10 +15,10 @@ class AvhrrNdviTest(unittest.TestCase):
     def test_process(self):
         dataset = xr.Dataset()
 
-        ch1_data = np.asarray([[1195, 1228], [902, 1747], [902, 1848]]) * 1e-4
+        ch1_data = np.asarray([[1195, 1228], [902, 1747], [902, 1848]], dtype=np.float64) * np.float64(1e-4)
         dataset["Ch1"] = Variable(["y", "x"], ch1_data)
 
-        ch2_data = np.asarray([[886, 946], [608, 1514], [596, 1634]]) * 1e-4
+        ch2_data = np.asarray([[886, 946], [608, 1514], [596, 1634]], dtype=np.float64) * np.float64(1e-4)
         dataset["Ch2"] = Variable(["y", "x"], ch2_data)
 
         result = self.avhrr_ndvi.process(dataset)
